Moved data retrieval to main component (spinner effect gets improved)
