+++ src/Components/Dashboard.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 
 import DataTable from "./DataTable";
 
@@ -6,211 +6,209 @@ import TouristArrivalsChartsContainer from "./TouristArrivalsChartsContainer";
 
 import SpendingChartsContainer from "./SpendingChartsContainer";
 
-import axios from "axios";
-
 const Dashboard = (props) => {
-  const [data, setData] = useState(null);
+  // const [data, setData] = useState(null);
 
-  useEffect(() => {
-    console.log("sectUseEffect", props.activeSection);
-    // `http://54.77.111.120:5300/data-grids/summary?token=${apiToken}`
-    const sectionIds = [
-      {
-        _id: "5fad0695e5fedf3df0e77be1",
-        section: "ecs_tourist_arrivals",
-        title: {
-          ca: "ARRIBADA DE TURISTES / AGOST 2020",
-        },
-      },
-      {
-        _id: "5fad0d1ae5fedf3df0e77be2",
-        section: "ecs_tourist_arrivals",
-        title: {
-          ca: "ARRIBADA DE TURISTES / GENER-AGOST 2020",
-        },
-      },
-      {
-        _id: "5fad0fbfe5fedf3df0e77be3",
-        section: "ecs_spending",
-        title: {
-          ca: "DESPESA DELS TURISTES / AGOST 2020  ",
-        },
-      },
-      {
-        _id: "5fad1be4e5fedf3df0e77be4",
-        section: "ecs_spending",
-        title: {
-          ca: "DESPESA DELS TURISTES / GENER-AGOST 2020",
-        },
-      },
-      {
-        _id: "5fad238de5fedf3df0e77be5",
-        section: "ecs_spending",
-        type: "total",
-        title: {
-          ca: "DESPESA DELS TURISTES / AGOST 2020  TOTAL",
-        },
-      },
-      {
-        _id: "5fad23abe5fedf3df0e77be6",
-        section: "ecs_spending",
-        type: "total",
-        title: {
-          ca: "DESPESA DELS TURISTES / GENER-AGOST 2020 TOTAL",
-        },
-      },
-      {
-        _id: "5fad23cde5fedf3df0e77be7",
-        section: "ecs_spending",
-        type: "stays",
-        title: {
-          ca: "PERNOCTACIONS I ESTADA MITJANA / AGOST 2020  ",
-        },
-      },
-      {
-        _id: "5fad23e8e5fedf3df0e77be8",
-        section: "ecs_spending",
-        type: "stays",
-        title: {
-          ca: "PERNOCTACIONS I ESTADA MITJANA / AGOST 2020 PER ILLA",
-        },
-      },
-      {
-        _id: "5fad2b76e5fedf3df0e77bec",
-        section: "ecs_spending",
-        type: "stays",
-        title: {
-          ca: "PERNOCTACIONS I ESTADA MITJANA / GENER - AGOST 2020  ",
-        },
-      },
-      {
-        _id: "5fad2bc0e5fedf3df0e77bed",
-        section: "ecs_spending",
-        type: "stays",
-        title: {
-          ca: "PERNOCTACIONS I ESTADA MITJANA / GENER-AGOST 2020 PER ILLA",
-        },
-      },
-      {
-        _id: "5fae7f9270ad955c4544c38e",
-        section: "ecs_occupancy",
-        title: {
-          ca: "OCUPACIÓ PER PLACES OBERTES / AGOST 2020",
-        },
-      },
-      {
-        _id: "5fae7f920dc77d5c4bfaf4ce",
-        section: "ecs_air_passengers_arrivals",
-        title: {
-          ca: "PASSATGERS AERIS (ARRIBADES) / AGOST 2020",
-        },
-      },
-      {
-        _id: "5fae7f9270ad955c4544c38f",
-        section: "ecs_occupancy",
-        title: {
-          ca: "OCUPACIÓ PER PLACES OBERTES / ACUMULAT 2020",
-        },
-      },
-      {
-        _id: "5fae7f920dc77d5c4bfaf4cf",
-        section: "ecs_air_passengers_arrivals",
-        title: {
-          ca: "PASSATGERS AERIS (ARRIBADES) / GENER-AGOST 2020",
-        },
-      },
+  // useEffect(() => {
+  //   console.log("sectUseEffect", props.activeSection);
+  //   // `http://54.77.111.120:5300/data-grids/summary?token=${apiToken}`
+  //   const sectionIds = [
+  //     {
+  //       _id: "5fad0695e5fedf3df0e77be1",
+  //       section: "ecs_tourist_arrivals",
+  //       title: {
+  //         ca: "ARRIBADA DE TURISTES / AGOST 2020",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad0d1ae5fedf3df0e77be2",
+  //       section: "ecs_tourist_arrivals",
+  //       title: {
+  //         ca: "ARRIBADA DE TURISTES / GENER-AGOST 2020",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad0fbfe5fedf3df0e77be3",
+  //       section: "ecs_spending",
+  //       title: {
+  //         ca: "DESPESA DELS TURISTES / AGOST 2020  ",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad1be4e5fedf3df0e77be4",
+  //       section: "ecs_spending",
+  //       title: {
+  //         ca: "DESPESA DELS TURISTES / GENER-AGOST 2020",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad238de5fedf3df0e77be5",
+  //       section: "ecs_spending",
+  //       type: "total",
+  //       title: {
+  //         ca: "DESPESA DELS TURISTES / AGOST 2020  TOTAL",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad23abe5fedf3df0e77be6",
+  //       section: "ecs_spending",
+  //       type: "total",
+  //       title: {
+  //         ca: "DESPESA DELS TURISTES / GENER-AGOST 2020 TOTAL",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad23cde5fedf3df0e77be7",
+  //       section: "ecs_spending",
+  //       type: "stays",
+  //       title: {
+  //         ca: "PERNOCTACIONS I ESTADA MITJANA / AGOST 2020  ",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad23e8e5fedf3df0e77be8",
+  //       section: "ecs_spending",
+  //       type: "stays",
+  //       title: {
+  //         ca: "PERNOCTACIONS I ESTADA MITJANA / AGOST 2020 PER ILLA",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad2b76e5fedf3df0e77bec",
+  //       section: "ecs_spending",
+  //       type: "stays",
+  //       title: {
+  //         ca: "PERNOCTACIONS I ESTADA MITJANA / GENER - AGOST 2020  ",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fad2bc0e5fedf3df0e77bed",
+  //       section: "ecs_spending",
+  //       type: "stays",
+  //       title: {
+  //         ca: "PERNOCTACIONS I ESTADA MITJANA / GENER-AGOST 2020 PER ILLA",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fae7f9270ad955c4544c38e",
+  //       section: "ecs_occupancy",
+  //       title: {
+  //         ca: "OCUPACIÓ PER PLACES OBERTES / AGOST 2020",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fae7f920dc77d5c4bfaf4ce",
+  //       section: "ecs_air_passengers_arrivals",
+  //       title: {
+  //         ca: "PASSATGERS AERIS (ARRIBADES) / AGOST 2020",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fae7f9270ad955c4544c38f",
+  //       section: "ecs_occupancy",
+  //       title: {
+  //         ca: "OCUPACIÓ PER PLACES OBERTES / ACUMULAT 2020",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fae7f920dc77d5c4bfaf4cf",
+  //       section: "ecs_air_passengers_arrivals",
+  //       title: {
+  //         ca: "PASSATGERS AERIS (ARRIBADES) / GENER-AGOST 2020",
+  //       },
+  //     },
 
-      {
-        _id: "5fb241b270ad955c4544c394",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca: " CREUERISTES / AGOST 2020 - Autoritat Portuària",
-        },
-      },
-      {
-        _id: "5fb3b9765d85e95533e27a92",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca: " CREUERISTES / AGOST 2020 - Ports IB",
-        },
-      },
-      {
-        _id: "5fb3b9760ab2ff553d13ebcc",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca: " CREUERISTES / GENER-AGOST 2020 - Autoritat Portuària",
-        },
-      },
-      {
-        _id: "5fb3b9760ab2ff553d13ebcd",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca: " CREUERISTES / GENER-AGOST 2020 - Ports IB",
-        },
-      },
-      {
-        _id: "5fb426715d85e95533e27a95",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca:
-            "PASSATGERS MARÍTIMS (ARRIBADES) /   AGOST 2020 - Autoritat Portuària",
-        },
-      },
-      {
-        _id: "5fb42ac70ab2ff553d13ebd0",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca: "PASSATGERS MARÍTIMS (ARRIBADES) /   AGOST 2020 - Ports IB",
-        },
-      },
-      {
-        _id: "5fb42b4f5d85e95533e27a96",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca:
-            "PASSATGERS MARÍTIMS (ARRIBADES) / GENER-AGOST 2020 - Autoritat Portuària",
-        },
-      },
-      {
-        _id: "5fb42b850ab2ff553d13ebd1",
-        section: "ecs_sea_passengers_arrivals",
-        title: {
-          ca: "PASSATGERS MARÍTIMS (ARRIBADES) / GENER-AGOST 2020 - Ports IB",
-        },
-      },
-    ];
+  //     {
+  //       _id: "5fb241b270ad955c4544c394",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca: " CREUERISTES / AGOST 2020 - Autoritat Portuària",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb3b9765d85e95533e27a92",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca: " CREUERISTES / AGOST 2020 - Ports IB",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb3b9760ab2ff553d13ebcc",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca: " CREUERISTES / GENER-AGOST 2020 - Autoritat Portuària",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb3b9760ab2ff553d13ebcd",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca: " CREUERISTES / GENER-AGOST 2020 - Ports IB",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb426715d85e95533e27a95",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca:
+  //           "PASSATGERS MARÍTIMS (ARRIBADES) /   AGOST 2020 - Autoritat Portuària",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb42ac70ab2ff553d13ebd0",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca: "PASSATGERS MARÍTIMS (ARRIBADES) /   AGOST 2020 - Ports IB",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb42b4f5d85e95533e27a96",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca:
+  //           "PASSATGERS MARÍTIMS (ARRIBADES) / GENER-AGOST 2020 - Autoritat Portuària",
+  //       },
+  //     },
+  //     {
+  //       _id: "5fb42b850ab2ff553d13ebd1",
+  //       section: "ecs_sea_passengers_arrivals",
+  //       title: {
+  //         ca: "PASSATGERS MARÍTIMS (ARRIBADES) / GENER-AGOST 2020 - Ports IB",
+  //       },
+  //     },
+  //   ];
 
-    let currentSectionIds = sectionIds.filter(
-      (item) => item.section === props.activeSection //TODO; change for active section (manage crash when navigating)
-    );
+  //   let currentSectionIds = sectionIds.filter(
+  //     (item) => item.section === props.activeSection //TODO; change for active section (manage crash when navigating)
+  //   );
 
-    // PUBLIC API TOKEN (does not expire. Cannot perform actions other than reading)
-    const apiToken =
-      "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImZyb250ZW5kIiwiaWF0IjoxNjA1NTI0OTkzfQ.E_70V52jXRF-dSNBhCawExXyvTnLuMljsWzjuWuSC9k";
-    const headers = {
-      "Content-Type": "application/json",
-    };
+  //   // PUBLIC API TOKEN (does not expire. Cannot perform actions other than reading)
+  //   const apiToken =
+  //     "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImZyb250ZW5kIiwiaWF0IjoxNjA1NTI0OTkzfQ.E_70V52jXRF-dSNBhCawExXyvTnLuMljsWzjuWuSC9k";
+  //   const headers = {
+  //     "Content-Type": "application/json",
+  //   };
 
-    let requests = currentSectionIds.map((sectionId) =>
-      axios.get(
-        `http://54.77.111.120:5300/data-grids/${sectionId._id}?token=${apiToken}`,
-        { headers: headers }
-      )
-    );
+  //   let requests = currentSectionIds.map((sectionId) =>
+  //     axios.get(
+  //       `http://54.77.111.120:5300/data-grids/${sectionId._id}?token=${apiToken}`,
+  //       { headers: headers }
+  //     )
+  //   );
 
-    axios
-      .all(requests)
-      .then(
-        axios.spread((...responses) => {
-          props.handleLoading();
-          setData(responses.map((resp) => resp.data));
-        })
-      )
+  //   axios
+  //     .all(requests)
+  //     .then(
+  //       axios.spread((...responses) => {
+  //         props.handleLoading();
+  //         setData(responses.map((resp) => resp.data));
+  //       })
+  //     )
 
-      .catch((error) => {
-        console.log("axios error", error);
-      });
-  }, [props.activeSection, props]);
+  //     .catch((error) => {
+  //       console.log("axios error", error);
+  //     });
+  // }, [props.activeSection, props]);
 
   const selectedView = (view) => {
     switch (view) {
@@ -242,13 +240,13 @@ const Dashboard = (props) => {
         };
         return (
           <div style={{ padding: 20 }}>
-            {data && (
+            {props.data && (
               <TouristArrivalsChartsContainer
                 data={touristArrivalChartsData}
               ></TouristArrivalsChartsContainer>
             )}
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -306,14 +304,14 @@ const Dashboard = (props) => {
         };
         return (
           <div style={{ padding: 20 }}>
-            {data && (
+            {props.data && (
               <SpendingChartsContainer
                 data={spendingChartsData}
               ></SpendingChartsContainer>
             )}
 
-            {data &&
-              data
+            {props.data &&
+              props.data
                 .slice(0, 2)
                 .map((tableInput, i) => (
                   <DataTable
@@ -327,8 +325,8 @@ const Dashboard = (props) => {
                   ></DataTable>
                 ))}
 
-            {data &&
-              data
+            {props.data &&
+              props.data
                 .slice(2, 4)
                 .map((tableInput, i) => (
                   <DataTable
@@ -340,8 +338,8 @@ const Dashboard = (props) => {
                     ]}
                   ></DataTable>
                 ))}
-            {data &&
-              data
+            {props.data &&
+              props.data
                 .slice(4)
                 .map((tableInput, i) => (
                   <DataTable key={i} data={tableInput} footer={[]}></DataTable>
@@ -354,8 +352,8 @@ const Dashboard = (props) => {
       case "ecs_occupancy":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -371,8 +369,8 @@ const Dashboard = (props) => {
       case "ecs_air_passengers_arrivals":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -389,8 +387,8 @@ const Dashboard = (props) => {
       case "ecs_sea_passengers_arrivals":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -403,8 +401,8 @@ const Dashboard = (props) => {
       case "ens_energy_demand":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -421,8 +419,8 @@ const Dashboard = (props) => {
       case "ens_human_pressure":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -439,8 +437,8 @@ const Dashboard = (props) => {
       case "sos_affiliates":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -457,8 +455,8 @@ const Dashboard = (props) => {
       case "sos_unemployed":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -475,8 +473,8 @@ const Dashboard = (props) => {
       case "sos_temporality":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
@@ -493,8 +491,8 @@ const Dashboard = (props) => {
       case "sos_companies":
         return (
           <div style={{ padding: 20 }}>
-            {data &&
-              data.map((tableInput, i) => (
+            {props.data &&
+              props.data.map((tableInput, i) => (
                 <DataTable
                   key={i}
                   data={tableInput}
